Migrating from jscience to Units of Measurement Systems
This is to move away from the jscience library which hasn't been updated since 2012 to the new JSR 363 compatible library.
see https://github.com/unitsofmeasurement/uom-systems

Somewhat of a work in progress as this breaks some of the UI when switching between unit types on the Output Volume field.
Need to map the R.strings to the values showin in the UnitMapping class.